refactor: split to-hit calculations by concern
