+++ src/utils/gameplay/toHit/index.ts
@@ -1,0 +1,38 @@
+export {
+  ATTACKER_MOVEMENT_MODIFIERS,
+  HEAT_THRESHOLDS,
+  PROBABILITY_TABLE,
+  RANGE_MODIFIERS,
+} from './constants';
+export {
+  aggregateModifiers,
+  formatToHitBreakdown,
+  getProbability,
+  simpleToHit,
+} from './aggregation';
+export {
+  calculateActuatorDamageModifier,
+  calculateAttackerMovementModifier,
+  calculateAttackerProneModifier,
+  calculateCalledShotModifier,
+  calculateHeatModifier,
+  calculateHullDownModifier,
+  calculateImmobileModifier,
+  calculateIndirectFireModifier,
+  calculateMinimumRangeModifier,
+  calculatePartialCoverModifier,
+  calculatePilotWoundModifier,
+  calculateProneModifier,
+  calculateRangeModifier,
+  calculateSecondaryTargetModifier,
+  calculateSensorDamageModifier,
+  calculateTargetingComputerModifier,
+  calculateTMM,
+  createBaseModifier,
+  getRangeBracket,
+  getRangeModifierForBracket,
+  getTerrainToHitModifier,
+} from './modifiers';
+export { calculateToHit, calculateToHitFromContext } from './calculate';
+export { calculateToHitWithC3 } from './c3';
+export type { IC3ToHitInput } from './c3';
